IAI_Backend/requirement.txt
fastapi
uvicorn[standard]
python-multipart
pydantic
PyPDF2
chromadb
sentence-transformers
transformers
torch
python-dotenv
requests
numpy
pandas
google-generativeai